requirements.txt
gunicorn==20.1.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
Flask==2.3.2
frozenlist==1.3.3
google-api-core==2.11.0
google-auth==2.19.1
google-cloud-core==2.3.2
google-cloud-datastore==2.15.2
google-cloud-logging==3.5.0
googleapis-common-protos==1.59.0
grpcio==1.54.2
grpcio-status==1.54.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
loguru==0.7.0
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.24.3
openai==0.27.8
pandas==2.0.2
pinecone-client==2.2.2
proto-plus==1.22.2
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
rsa==4.9
six==1.16.0
tiktoken==0.4.0
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.6
win32-setctime==1.1.0
yarl==1.9.2